requirements.txt
flask==3.0.2
python-telegram-bot==20.8
google-generativeai==0.3.2
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
reportlab==4.1.0
openai==1.12.0
openai-agents==0.1.0
aiohttp==3.9.3
asyncio==3.4.3
typing-extensions==4.9.0
pydantic==2.6.1 